refactor(analytics): consolidate analytics utilities into lib/analytics/
- Create lib/analytics/ with hooks.ts, utils.ts, and shared.tsx
- Move PostHogPageTracker (was PostHogPageView), WebVitals, and
  CustomAnalyticsScript into shared.tsx; PostHogPageTracker now
  owns its own Suspense boundary
- Move analytics event tracking (track, AnalyticsEvent, etc.) from
  lib/analytics.ts into lib/analytics/utils.ts
- Move usePHFeatureFlag (was useFeatureFlag) and PHFeatureFlags
  registry (was FEATURE_FLAGS in lib/featureFlags.ts) into
  lib/analytics/hooks.ts
- Switch LanguageModelsPage LLM config gate from env-var constant
  to PostHog feature flag via usePHFeatureFlag
