+++ web/src/sections/modals/languageModels/AnthropicModal.tsx
@@ -10,7 +10,7 @@ import {
   buildValidationSchema,
 } from "@/sections/modals/languageModels/utils";
 import { submitProvider } from "@/sections/modals/languageModels/svc";
-import { LLMProviderConfiguredSource } from "@/lib/analytics";
+import { LLMProviderConfiguredSource } from "@/lib/analytics/utils";
 import {
   APIKeyField,
   ModelSelectionField,
