+++ web/src/sections/modals/languageModels/CustomModal.tsx
@@ -11,7 +11,7 @@ import type { ModelConfiguration } from "@/lib/languageModels/types";
 import * as Yup from "yup";
 import { useInitialValues } from "@/sections/modals/languageModels/utils";
 import { submitProvider } from "@/sections/modals/languageModels/svc";
-import { LLMProviderConfiguredSource } from "@/lib/analytics";
+import { LLMProviderConfiguredSource } from "@/lib/analytics/utils";
 import {
   APIKeyField,
   APIBaseField,
