+++ web/src/sections/modals/languageModels/OllamaModal.tsx
@@ -19,7 +19,7 @@ import {
   mergeFetchedModelConfigurations,
 } from "@/sections/modals/languageModels/utils";
 import { submitProvider } from "@/sections/modals/languageModels/svc";
-import { LLMProviderConfiguredSource } from "@/lib/analytics";
+import { LLMProviderConfiguredSource } from "@/lib/analytics/utils";
 import {
   CONTAINERIZED_HOST_NOTE,
   ModelSelectionField,
